+++ docker/conf/phpmyadmin/config.user.inc.php
@@ -1,3 +1,0 @@
-<?php
-
-$cfg['MaxNavigationItems'] = 500;