+++ src/Lib/WebSocketConnection.php
@@ -11,8 +11,8 @@ use function Sentry\captureException;
  * This class is for handling connections to our WebSocket server.
  *
  * Because it has to handle many connections at the same time, the server written in JavaScript/Node.js and not in PHP.
- * For historical reasons, the docker container containing our WebSocket server has been called "chat". Its API is
- * provided by chat/src/RestController.ts.
+ * For historical reasons, the docker container containing our WebSocket server has been called "websocket". Its API is
+ * provided by websocket/src/RestController.ts.
  */
 class WebSocketConnection
 {